# service/src/main/resources/api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: realmeet
servers:
  - url: 'http://localhost:3000'
paths:
  /rooms/{id}:
    get:
      summary: 'Return a room by ID'
      operationId: getRoom
      tags:
        - room
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the room
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDTO'
        404:
          description: Resource Not Found

    delete:
      summary: 'Delete a room by ID'
      operationId: deleteRoom
      tags:
        - room
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the room
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Deleted Successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDTO'
        404:
          description: Resource Not Found
  /rooms:
    post:
      summary: "Creates a room"
      operationId: createRoom
      tags:
        - room
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomDTO'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDTO'
        422:
          $ref: '#/components/responses/UnprocessableEntity'

components:
  schemas:
    CreateRoomDTO:
      type: object
      properties:
        name:
          type: string
        seats:
          type: integer

    RoomDTO:
      type: object
      allOf:
        - $ref: '#/components/schemas/CreateRoomDTO'
        - type: object
          properties:
            id:
              type: integer
              format: int64

    ResponseError:
      type: object
      required:
        - field
        - errorCode
      properties:
        field:
          type: string
        errorCode:
          type: string

  responses:
    UnprocessableEntity:
      description: The server understands the content type of the request entity, but was unable to process the contained instructions.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ResponseError"